2.5.0
2.5.0

+++ BiostarMobile/BiostarMobile/Classes/Model/Card/Card.m
@@ -24,43 +24,10 @@
 {
     NSString *description;
     
-    description = [NSString stringWithFormat:@"%ld", self.fingerprint_templates.count];
+    description = [NSString stringWithFormat:@"%ld", (unsigned long)self.fingerprint_templates.count];
     
     return description;
     
-//    NSString *description;
-//    
-//    NSUInteger index = [self.fingerprint_index_list[0] integerValue];
-//    switch (index)
-//    {
-//        case 0:
-//            description = NSLocalizedString(@"1st_fingerprint", nil);
-//            break;
-//            
-//        case 1:
-//            description = NSLocalizedString(@"2nd_fingerprint", nil);
-//            break;
-//            
-//        case 2:
-//            description = NSLocalizedString(@"3rd_fingerprint", nil);
-//            break;
-//            
-//        default:
-//            
-//            description = [NSString stringWithFormat:NSLocalizedString(@"%ldth_fingerprint", nil), (long)index + 1];
-//            break;
-//    }
-//    
-//    if (self.fingerprint_index_list.count > 1)
-//    {
-//        NSMutableString *extraDec = [[NSMutableString alloc] init];
-//        
-//        [extraDec appendFormat:@" +%ld",self.fingerprint_index_list.count - 1];
-//        
-//        description = [NSString stringWithFormat:@"%@%@", description,  extraDec];
-//    }
-//    
-//    return description;
 }
